Scop: extract activePromotionsIndexes from activePromotions
Internally, we may need to modify the stored active promotions but the
public functions return either a copy of or a const reference to that
storage.  Extract logic to find active promotions into a separate
function that returns indexes into the storage, and use it to create a
copy inside a public call.